style(controller): 클래스명 수정
PlayerCharacterController -> ActionController 이름 변경
CharacterAnimation 클래스의 디버그 로그 제거
