+++ Assets/02. Scripts/Character/Controller/ControllerComponent.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 94a5668a319560544894526c1ae214f4
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
