+++ Assets/09. Animations/Animator/AC_PlayerCharacter.controller
@@ -368,7 +368,7 @@ AnimatorStateTransition:
   m_TransitionDuration: 0.9
   m_TransitionOffset: 0
   m_ExitTime: 0.9
-  m_HasExitTime: 0
+  m_HasExitTime: 1
   m_HasFixedDuration: 0
   m_InterruptionSource: 0
   m_OrderedInterruption: 1
